More project refactoring
